hms_backend/requirements.txt
Django==4.2.17
djangorestframework==3.14.0
djangorestframework-simplejwt==5.4.0
mongoengine==0.29.1
pymongo==4.10.1
django-cors-headers==4.6.0
python-decouple==3.8
PyJWT==2.10.1
